perf: extract navbar mobile
